requirements.txt
future==0.18.2
joblib==0.17.0
numpy==1.19.2
Pillow==7.2.0
scikit-learn==0.23.2
scipy==1.5.2
sklearn==0.0
threadpoolctl==2.1.0
torch==1.6.0
torchvision==0.7.0
cycler==0.10.0
decorator==4.4.2
imageio==2.9.0
joblib==1.0.0
kiwisolver==1.3.1
matplotlib==3.3.4
networkx==2.5
numpy==1.20.0
opencv-python==4.5.1.48
pandas==1.2.1
Pillow==8.1.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyWavelets==1.1.1
scikit-image==0.18.1
scikit-learn==0.24.1
scipy==1.6.0
six==1.15.0
threadpoolctl==2.1.0
tifffile==2021.1.14
torch==1.7.1
torchvision==0.8.2
tqdm==4.56.0
typing-extensions==3.7.4.3
Wand==0.6.5
